+++ addons/RatingAddon/RatingAddon.cs
@@ -46,9 +46,11 @@ namespace RatingAddon
 
         void mpv_ClientMessage(string[] args)
         {
-            int rating;
+            if (args == null ||
+                args.Length != 2 ||
+                args[0] != "rate-file" ||
+                !int.TryParse(args[1], out int rating))
 
-            if (args?.Length != 2 || args[0] != "rate-file" || ! int.TryParse(args[1], out rating))
                 return;
 
             Dic[mp.get_property_string("path")] = rating;