+++ mpv.net/WinForms/MainForm.cs
@@ -515,6 +515,8 @@ namespace mpvnet
         {
             base.OnActivated(e);
             CheckUrlInClipboard();
+            Message m = new Message() { Msg = 0x0202 }; // WM_LBUTTONUP
+            Native.SendMessage(Handle, m.Msg, m.WParam, m.LParam);
         }
 
         void CheckUrlInClipboard()